Refactor: PostButton 공통 컴퍼넌트로 만들기 및 전체보기 버튼 간격수정
